backend/ai-service/requirements.txt
# Core dependencies - Pydantic v2 호환 버전
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6  # Required for FastAPI file uploads

# Flask as fallback option
flask==2.3.3

# AI/ML Core
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0  # Required for time series analysis

# LLM/NLP
httpx==0.25.2  # CRITICAL: Must install this BEFORE openai
openai==1.35.0  # Compatible with httpx 0.25.2
google-generativeai==0.3.2
transformers==4.36.2
sentence-transformers==2.2.2

# Audio processing
librosa==0.10.2
resampy==0.4.3
soundfile==0.12.1
pydub==0.25.1

# Deep Learning (SincNet)
# torch==2.1.2+cpu  # CPU-only version - installed separately in Dockerfile
# torchaudio==2.1.2+cpu
# torchvision==0.16.2+cpu

# Google Cloud
google-cloud-speech==2.23.0
google-cloud-storage==2.10.0
firebase-admin==6.1.0

# Vector store / RAG
chromadb==0.4.22
langchain==0.1.0
faiss-cpu==1.7.4

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0